Model emprestimo feito

+++ database/migrations/2024_10_03_124424_create_emprestimos_table.php
@@ -16,7 +16,7 @@ class CreateEmprestimosTable extends Migration
         Schema::create('emprestimos', function (Blueprint $table) {
             $table->id();
             $table->text("descricao");
-            $table->integer("quantia");
+            $table->decimal("quantia", 15, 2);
             $table->unsignedBigInteger("id_usuario");
             $table->foreign("id_usuario")->references("id")->on("users")->onDelete("cascade");
             $table->timestamps();
